# backend/fastapi/openapi-fastapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workma!te AI API",
    "description": "회의체 운영 AI Agent 서비스의 **AI · 실시간 · 지식그래프** 백엔드 (FastAPI).\n\n- **인증**: Spring이 발급한 JWT access token을 `Authorize`(HTTPBearer)에 입력하세요. (`JWT_SECRET`을 Spring과 공유, HS256)\n- **스트리밍**: 다수 엔드포인트가 SSE(`text/event-stream`)로 토큰 단위 응답합니다.\n- **데이터 스코프**: AI 도구는 사용자의 소속 회의체로 접근을 제한합니다(IDOR 이중 방어).\n- 인증·도메인 CRUD는 별도 Spring Boot 서버(`/api/v1`)가 담당합니다.",
    "contact": {
      "name": "Team No.9"
    },
    "license": {
      "name": "Internal"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agent/hitl-reviews": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "HITL 검토 생성",
        "description": "보고서/아젠다에 대한 사람(HITL) 검토 결과를 저장하고 대상 노드에 흡수 동기화한다.\n\n대상 회의체의 require_meeting_edit 가드(간사/회사관리자/시스템관리자)가 필요하다.",
        "operationId": "create_hitl_review_api_agent_hitl_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agent/hitl-reviews/{hj_id}": {
      "patch": {
        "tags": [
          "agents"
        ],
        "summary": "HITL 검토 수정",
        "description": "기존 HITL 검토의 상태·코멘트를 수정하고 대상 노드에 다시 흡수 동기화한다.\n\n대상 회의체의 require_meeting_edit 가드(간사/회사관리자/시스템관리자)가 필요하다.",
        "operationId": "update_hitl_review_api_agent_hitl_reviews__hj_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hj_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Hj Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlReviewPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/reports/global-review/stream": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "보고서 종합 검토 — SSE 스트리밍",
        "description": "회의체의 여러 보고서를 종합 검토한 결과를 SSE(text/event-stream)로 스트리밍한다.\n\nrequire_view 가드로 회의체 열람 권한을 확인한다.",
        "operationId": "global_review_stream_ep_api_agent_reports_global_review_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agent/reports/review": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "보고서 단건 검토",
        "description": "단일 보고서 내용을 안건 기준으로 검토해 점수·피드백을 반환한다.\n\n로그인 사용자(get_current_user) 인증이 필요하며, 검토 생성 실패 시 502를 반환한다.",
        "operationId": "review_report_ep_api_agent_reports_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agent/reports/review/start": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "보고서 HITL 검토 시작",
        "description": "보고서 HITL 검토 워크플로우를 시작하고 서버 발급 thread_id(run_id)를 반환한다.\n\n로그인 사용자(get_current_user) 인증이 필요하다.",
        "operationId": "start_report_review_ep_api_agent_reports_review_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartReportReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agent/reports/review/confirm": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "보고서 HITL 검토 확정",
        "description": "보고서 HITL 검토를 승인/반려로 확정하고 결과를 저장한다.\n\nmeeting_id가 있으면 require_meeting_edit 가드를 적용하고, 멱등성 가드로\n더블클릭·재시도 중복을 차단한다.",
        "operationId": "confirm_report_review_ep_api_agent_reports_review_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmReportReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agent/extraction/review/start": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "과제 추출 HITL 검토 시작",
        "description": "과제(아젠다) 추출 HITL 검토를 시작하고 서버 발급 thread_id(run_id)를 반환한다.\n\n로그인 사용자(get_current_user) 인증이 필요하다.",
        "operationId": "start_extraction_review_ep_api_agent_extraction_review_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agent/extraction/review/confirm": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "과제 추출 HITL 검토 확정",
        "description": "과제(아젠다) 추출 HITL 검토를 승인/반려로 확정하고 결과를 저장한다.\n\nmeeting_id가 있으면 require_meeting_edit 가드를 적용하고, 멱등성 가드로\n더블클릭·재시도 중복을 차단한다.",
        "operationId": "confirm_extraction_review_ep_api_agent_extraction_review_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmExtractionRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "meeting_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Id"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "approved"
        ],
        "title": "ConfirmExtractionRequest"
      },
      "ConfirmReportReviewRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": ""
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "default": ""
          },
          "meeting_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Id"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "approved"
        ],
        "title": "ConfirmReportReviewRequest"
      },
...
      "GlobalReviewRequest": {
        "properties": {
          "meeting_id": {
            "type": "integer",
            "title": "Meeting Id"
          },
          "reports_info": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Reports Info"
          },
          "chat_history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat History",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "reports_info"
        ],
        "title": "GlobalReviewRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HitlReviewCreate": {
        "properties": {
          "target_type": {
            "type": "string",
            "title": "Target Type"
          },
          "agenda_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda Id"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "agent_log_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Log Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "edited"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "target_type"
        ],
        "title": "HitlReviewCreate"
      },
      "HitlReviewPatch": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "HitlReviewPatch"
      },
...
      "ReviewReportRequest": {
        "properties": {
          "report_content": {
            "type": "string",
            "title": "Report Content"
          },
          "agenda": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "report_content"
        ],
        "title": "ReviewReportRequest"
      },
      "StartExtractionRequest": {
        "properties": {
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "org_dept_list": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Dept List",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "StartExtractionRequest"
      },
      "StartReportReviewRequest": {
        "properties": {
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "report_content": {
            "type": "string",
            "title": "Report Content"
          },
          "agenda": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agenda",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "report_content"
        ],
        "title": "StartReportReviewRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}